user profile ui created

+++ MVC/CI-Platform/CI-Platform.Entities/ViewModels/CityVM.cs
@@ -2,7 +2,7 @@ namespace CI_Platform.Entities.ViewModels;
 
 public class CityVM
 {
-    public long CityId{get; set;}
+    public int CityId {get; set;}
 
     public string CityName{get; set;} = string.Empty;
