+++ MVC/CI-Platform/CI-Platform.Entities/ViewModels/CountryVM.cs
@@ -2,7 +2,7 @@ namespace CI_Platform.Entities.ViewModels;
 
 public class CountryVM
 {
-    public long CountryId { get; set; }
+    public short CountryId { get; set; }
 
     public string CountryName { get; set; } = string.Empty;
 }
